Starting The Feature: "Rewrite Front-End to React"
This is going to be a semi-long process. Right now the application renders EJS files on the server side. I want the shopping application to now render/operate on the client side. The Express app will continue to run, but now it will TRUELY be an API. As in, it will only respond to request from the client application which I'm going to rewrite into React

I started making this change by running "create-react-app". I figured out that I'm going to have two node_module packages now. I can start depreciating the jQuery files I was writing earlier... I wanted to rewrite that spaghetti code into a much cleaner application but this is a more practical approach.