I've decided that I need to decouple the client and server in the project structure completely. Things were getting too crossed with two node_modules folders and the IDE was getting pissed. I'm going to go back to the basics now...
I started out grinding HTML, CSS, and JS/jQuery. Now I'm going to go back and do as much of the front end as I can before I need to implement the backend (which is already done and moved to a separate file location). I'll be focusing on React completely for the next couple weeks or so.