I learned a lot more about React-Router. I still have a couple more videos to watch in order to make sure I completely understand the concept. I understand the libraries Route, Router, and Navigation components now though.
In the application I fixed the routing so that the pages in DOM are located in a <div class='container'>, which is in the <body> html element. Before, they were rendering from the <header> element. I also fixed some formatting and changed folder names so they're lowercase, and file names so they start with an Uppercase, indicating the file is for a component.