build(Client Side): Added Bulma and global SASS variables. I completed the form and believe I can make the second one much faster.
I added Bulma to the project after fidgetting around with "Ant Design"
for awhile. I decided that extending the SASS and incorporating it into
the components was much more beneficial than using pre-made ones. The
configuration for Ant Design's Less and the strictness of it put me off.
I still want to be more hands on with React for this project.

BREAKING CHANGE: CSS files