feat(Account Navigation): Added account naviation sidebar and columns
The account page is now split between the account navigation and the
panel informations. I have added the Account Information which needs to
be refactored, and the Shipping information which needs to be
refactored. I haven't yet added payment information or order
information. I will work on those soon. Using Bulma is increasing my
speed drastically, but the process is still slow.

BREAKING CHANGE: None

Closes Account Nav