feat(Account Page): Added order panel and payments panel to the account page
I made some changes to fix the layout of the account panels for a logged
in user. I added the order history and payment history for the user. I
need to work on adding an overall state that tracks if the user is
logged in or not.

BREAKING CHANGE: none

Closes none