feat(Frontend): Working on Checkout Components, redux
I finished most of the redux and crude logic for the shopping cart and
products...
I had a lot of insight about javascript functions and
function-objects, passing the
function down props, etc... I also started
using the react-redux hooks. I'm working on
making the checkout
components now, as It's one of the last things I need to do. After
I'm
done with that I can move on to using thunk and an API

BREAKING CHANGE: None

Closes None