feat(Frontend): Checkout Page
I added a lot of stuff to the react / redux frontend. The checkout page
is mostly complete. The order panel gets the orders from the cart store.
I did a crude implementation of adding a new payment card to the
"payment method" panel. I think I'm about done with all of this now...
now I need to run a backend API and then add async http methods to the
frontend.

BREAKING CHANGE: None

Closes None